Fixes suggested by Scrutinizer

+++ src/BusinessTime.php
@@ -75,7 +75,7 @@ class BusinessTime extends Carbon
         // problem" that Monday 09:00 + 1 business day could technically be
         // Monday 17:00, but intuitively should be Tuesday 09:00.
         $daysToJump = (int) $businessDaysToAdd;
-        $next = $this->copy()->addDays($daysToJump);
+        $next       = $this->copy()->addDays($daysToJump);
 
         // We need to check how much business time we actually covered by
         // skipping ahead in days.
@@ -130,7 +130,7 @@ class BusinessTime extends Carbon
         // problem" that Tuesday 17:00 - 1 business day could technically be
         // Tuesday 09:00, but intuitively should be Monday 17:00.
         $daysToJump = (int) $businessDaysToSub;
-        $prev = $this->copy()->subDays($daysToJump);
+        $prev       = $this->copy()->subDays($daysToJump);
 
         // We need to check how much business time we actually covered by
         // skipping back in days.
@@ -180,7 +180,7 @@ class BusinessTime extends Carbon
             return $this->subBusinessHours($businessHoursToAdd * -1);
         }
 
-        $next = $this->copy();
+        $next      = $this->copy();
         $decrement = $this->precision()->inHours();
         while ($businessHoursToAdd > 0) {
             if ($next->isBusinessTime()) {
@@ -223,7 +223,7 @@ class BusinessTime extends Carbon
             return $this->addBusinessHours($businessHoursToSub * -1);
         }
 
-        $prev = $this->copy();
+        $prev      = $this->copy();
         $decrement = $this->precision()->inHours();
         while ($businessHoursToSub > 0) {
             $prev = $prev->sub($this->precision());
@@ -382,10 +382,10 @@ class BusinessTime extends Carbon
     public function floor(?DateInterval $precision = null): self
     {
         $seconds = Interval::instance($precision ?: $this->precision())
-                            ->inSeconds();
+            ->inSeconds();
 
         return $this->copy()->setTimestamp(
-            floor($this->timestamp / $seconds) * $seconds
+            (int) (floor($this->timestamp / $seconds) * $seconds)
         );
     }
 
@@ -401,10 +401,10 @@ class BusinessTime extends Carbon
     public function round(?DateInterval $precision = null): self
     {
         $seconds = Interval::instance($precision ?: $this->precision())
-                            ->inSeconds();
+            ->inSeconds();
 
         return $this->copy()->setTimestamp(
-            round($this->timestamp / $seconds) * $seconds
+            (int) (round($this->timestamp / $seconds) * $seconds)
         );
     }
 
@@ -423,10 +423,10 @@ class BusinessTime extends Carbon
     public function ceil(?DateInterval $precision = null): self
     {
         $seconds = Interval::instance($precision ?: $this->precision())
-                            ->inSeconds();
+            ->inSeconds();
 
         return $this->copy()->setTimestamp(
-            ceil($this->timestamp / $seconds) * $seconds
+            (int) (ceil($this->timestamp / $seconds) * $seconds)
         );
     }
 
@@ -497,8 +497,8 @@ ERR
 
         return $this->setLengthOfBusinessDay(
             $this->copy()
-                    ->setTimestamp($typicalDay->startOfDay()->getTimestamp())
-                    ->diffBusiness($typicalDay->endOfDay())
+                ->setTimestamp($typicalDay->startOfDay()->getTimestamp())
+                ->diffBusiness($typicalDay->endOfDay())
         );
     }
 
@@ -594,13 +594,14 @@ ERR
         // We're taking a basic approach with some variables and a loop here as
         // it turns out to be ~25% faster than using Carbon::diffFiltered().
 
+        /** @var BusinessTime $start */
         $start = $this;
-        $end = $time;
-        $sign = 1;
+        $end   = $time;
+        $sign  = 1;
         // Swap if we're diffing back in time.
         if ($this > $time) {
             $start = $time;
-            $end = $this;
+            $end   = $this;
             // We only need to negate if absolute is false.
             $sign = $absolute ? 1 : -1;
         }
@@ -608,6 +609,7 @@ ERR
         // Count the business time diff by iterating in steps the length of the
         // precision and checking if each step counts as business time.
         $diff = 0;
+        /** @var BusinessTime $next */
         $next = $start->copy();
         while ($next < $end) {
             if ($next->isBusinessTime()) {
